+++ ipa-core/src/cli/config_parse.rs
@@ -359,5 +359,350 @@ pub fn sharded_server_from_toml_str(
 
 #[cfg(test)]
 mod tests {
+    use crate::cli::{config_parse::{parse_sharded_network_toml, Error}, sharded_server_from_toml_str};
 
+    use once_cell::sync::Lazy;
+    use crate::{
+        config::HttpClientConfigurator,
+        helpers::HelperIdentity,
+        sharding::ShardIndex,
+    };
+
+    #[test]
+    fn parse_sharded_server_happy() {
+        // Asuming position of the second helper in the second shard (the middle server in the 3 x 3)
+        let (mpc, shard) = sharded_server_from_toml_str(
+            &SHARDED_OK_REPEAT,
+            HelperIdentity::TWO,
+            ShardIndex::from(1),
+            ShardIndex::from(3),
+            None,
+        )
+        .unwrap();
+        assert_eq!(
+            vec![
+                "helper1.shard1.org:443",
+                "helper2.shard1.org:443",
+                "helper3.shard1.org:443"
+            ],
+            mpc.peers
+                .into_iter()
+                .map(|p| p.url.to_string())
+                .collect::<Vec<_>>()
+        );
+        assert_eq!(
+            vec![
+                "helper2.shard0.org:555",
+                "helper2.shard1.org:555",
+                "helper2.shard2.org:555"
+            ],
+            shard
+                .peers
+                .into_iter()
+                .map(|p| p.url.to_string())
+                .collect::<Vec<_>>()
+        );
+    }
+
+    /// Tests that the url of a shard gets updated with the shard url.
+    #[test]
+    fn transform_sharded_peers() {
+        let mut n = parse_sharded_network_toml(&SHARDED_OK_REPEAT).unwrap();
+        assert_eq!(
+            "helper3.shard2.org:666",
+            n.peers.pop().unwrap().to_shard_peer().url
+        );
+        assert_eq!(
+            "helper2.shard2.org:555",
+            n.peers.pop().unwrap().to_shard_peer().url
+        );
+    }
+
+    /// Expects an error if the number of peers isn't a multiple of 3
+    #[test]
+    fn invalid_nr_of_peers() {
+        assert!(matches!(
+            parse_sharded_network_toml(&SHARDED_8),
+            Err(Error::InvalidNetworkSize(_))
+        ));
+    }
+
+    /// If any sharded url is set (indicating this is a sharding config), then ALL urls must be set.
+    #[test]
+    fn parse_network_toml_shard_urls_some_set() {
+        assert!(matches!(
+            parse_sharded_network_toml(&SHARDED_COMPAT_ONE_URL),
+            Err(Error::MissingShardUrls(_))
+        ));
+    }
+
+    /// If there are more than 3 peers configured (indicating this is a sharding config), then ALL urls must be set.
+    #[test]
+    fn parse_network_toml_shard_urls_set() {
+        assert!(matches!(
+            parse_sharded_network_toml(&SHARDED_MISSING_URLS_REPEAT),
+            Err(Error::MissingShardUrls(_))
+        ));
+    }
+
+    /// Check that [`sharded_server_from_toml_str`] can work in the previous format.
+    #[test]
+    fn parse_sharded_without_shard_urls() {
+        let (mpc, mut shard) = sharded_server_from_toml_str(
+            &NON_SHARDED_COMPAT,
+            HelperIdentity::ONE,
+            ShardIndex::FIRST,
+            ShardIndex::from(1),
+            Some(666),
+        )
+        .unwrap();
+        assert_eq!(1, shard.peers.len());
+        assert_eq!(3, mpc.peers.len());
+        assert_eq!(
+            "helper1.org:666",
+            shard.peers.pop().unwrap().url.to_string()
+        );
+    }
+
+    /// Check that [`sharded_server_from_toml_str`] can work in the previous format, even when the
+    /// given MPC URL doesn't have a port (NOTE: helper 2 doesn't specify it).
+    #[test]
+    fn parse_sharded_without_shard_urls_no_port() {
+        let (mpc, mut shard) = sharded_server_from_toml_str(
+            &NON_SHARDED_COMPAT,
+            HelperIdentity::TWO,
+            ShardIndex::FIRST,
+            ShardIndex::from(1),
+            Some(666),
+        )
+        .unwrap();
+        assert_eq!(1, shard.peers.len());
+        assert_eq!(3, mpc.peers.len());
+        assert_eq!(
+            "helper2.org:666",
+            shard.peers.pop().unwrap().url.to_string()
+        );
+    }
+
+    /// Testing happy case of a sharded network config
+    #[test]
+    fn happy_parse_sharded_network_toml() {
+        let r_entire_network = parse_sharded_network_toml(SHARDED_OK);
+        assert!(r_entire_network.is_ok());
+        let entire_network = r_entire_network.unwrap();
+        assert!(matches!(
+            entire_network.client.http_config,
+            HttpClientConfigurator::Http2(_)
+        ));
+        assert_eq!(3, entire_network.peers.len());
+        assert_eq!("helper3.shard0.org:443", entire_network.peers[2].config.url);
+        assert_eq!(
+            "helper3.shard0.org:666",
+            entire_network.peers[2]
+                .shard_url
+                .as_ref()
+                .unwrap()
+                .to_string()
+        );
+    }
+
+    /// Testing happy case of a longer sharded network config
+    #[test]
+    fn happy_parse_larger_sharded_network_toml() {
+        let r_entire_network = parse_sharded_network_toml(&SHARDED_OK_REPEAT);
+        assert!(r_entire_network.is_ok());
+        let entire_network = r_entire_network.unwrap();
+        assert_eq!(9, entire_network.peers.len());
+        assert_eq!(
+            "helper3.shard2.org:666",
+            entire_network.peers[8]
+                .shard_url
+                .as_ref()
+                .unwrap()
+                .to_string()
+        );
+    }
+
+    /// This test validates that the new logic that handles sharded configurations can also handle the previous version
+    #[test]
+    fn parse_non_sharded_network_toml() {
+        let r_entire_network = parse_sharded_network_toml(&NON_SHARDED_COMPAT);
+        assert!(r_entire_network.is_ok());
+        let entire_network = r_entire_network.unwrap();
+        assert!(matches!(
+            entire_network.client.http_config,
+            HttpClientConfigurator::Http2(_)
+        ));
+        assert_eq!(3, entire_network.peers.len());
+        assert_eq!("helper3.org:443", entire_network.peers[2].config.url);
+    }
+
+    // Following are some large &str const used for tests
+
+    /// Valid: A non-sharded network toml, just how they used to be
+    static NON_SHARDED_COMPAT: Lazy<String> = Lazy::new(|| format!("{CLIENT}{P1}{REST}"));
+
+    /// Invalid: Same as [`NON_SHARDED_COMPAT`] but with a single `shard_port` set.
+    static SHARDED_COMPAT_ONE_URL: Lazy<String> =
+        Lazy::new(|| format!("{CLIENT}{P1}\nshard_url = \"helper1.org:777\"\n{REST}"));
+
+    /// Helper const used to create client configs
+    const CLIENT: &str = r#"[client.http_config]
+ping_interval_secs = 90.0
+version = "http2"
+"#;
+
+    /// Helper const that has the first part of a Peer, just before were `shard_port` should be
+    /// specified.
+    const P1: &str = r#"
+[[peers]]
+certificate = """
+-----BEGIN CERTIFICATE-----
+MIIBmzCCAUGgAwIBAgIIMlnveFys5QUwCgYIKoZIzj0EAwIwJjEkMCIGA1UEAwwb
+aGVscGVyMS5wcm9kLmlwYS1oZWxwZXIuZGV2MB4XDTI0MDkwNDAzMzMwM1oXDTI0
+MTIwNDAzMzMwM1owJjEkMCIGA1UEAwwbaGVscGVyMS5wcm9kLmlwYS1oZWxwZXIu
+ZGV2MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEWmrrkaKM7HQ0Y3ZGJtHB7vfG
+cT/hDCXCoob4pJ/fpPDMrqhiwTTck3bNOuzv9QIx+p5C2Qp8u67rYfK78w86NaNZ
+MFcwJgYDVR0RBB8wHYIbaGVscGVyMS5wcm9kLmlwYS1oZWxwZXIuZGV2MA4GA1Ud
+DwEB/wQEAwICpDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwCgYIKoZI
+zj0EAwIDSAAwRQIhAKVdDCQeXLRXDYXy4b1N1UxD/JPuD9H7zeRb8/nmIDTfAiBL
+a6L0t1Ug8i2RcequSo21x319Tvs5nUbGwzMFSS5wKA==
+-----END CERTIFICATE-----
+"""
+url = "helper1.org:443""#;
+
+    /// The rest of a configuration
+    /// Note the second helper doesn't provide a port as part of its url
+    const REST: &str = r#"
+[peers.hpke]
+public_key = "f458d5e1989b2b8f5dacd4143276aa81eaacf7449744ab1251ff667c43550756"
+
+[[peers]]
+certificate = """
+-----BEGIN CERTIFICATE-----
+MIIBmzCCAUGgAwIBAgIITOtoca16QckwCgYIKoZIzj0EAwIwJjEkMCIGA1UEAwwb
+aGVscGVyMi5wcm9kLmlwYS1oZWxwZXIuZGV2MB4XDTI0MDkwNDAzMzMwOFoXDTI0
+MTIwNDAzMzMwOFowJjEkMCIGA1UEAwwbaGVscGVyMi5wcm9kLmlwYS1oZWxwZXIu
+ZGV2MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAETxOH4ATz6kBxLuRznKDFRugm
+XKmH7mzRB9wn5vaVlVpDzf4nDHJ+TTzSS6Lb3YLsA7jrXDx+W7xPLGow1+9FNqNZ
+MFcwJgYDVR0RBB8wHYIbaGVscGVyMi5wcm9kLmlwYS1oZWxwZXIuZGV2MA4GA1Ud
+DwEB/wQEAwICpDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwCgYIKoZI
+zj0EAwIDSAAwRQIhAI4G5ICVm+v5KK5Y8WVetThtNCXGykUBAM1eE973FBOUAiAS
+XXgJe9q9hAfHf0puZbv0j0tGY3BiqCkJJaLvK7ba+g==
+-----END CERTIFICATE-----
+"""
+url = "helper2.org"
+
+[peers.hpke]
+public_key = "62357179868e5594372b801ddf282c8523806a868a2bff2685f66aa05ffd6c22"
+
+[[peers]]
+certificate = """
+-----BEGIN CERTIFICATE-----
+MIIBmzCCAUGgAwIBAgIIaf7eDCnXh2swCgYIKoZIzj0EAwIwJjEkMCIGA1UEAwwb
+aGVscGVyMy5wcm9kLmlwYS1oZWxwZXIuZGV2MB4XDTI0MDkwNDAzMzMxMloXDTI0
+MTIwNDAzMzMxMlowJjEkMCIGA1UEAwwbaGVscGVyMy5wcm9kLmlwYS1oZWxwZXIu
+ZGV2MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEIMqxCCtu4joFr8YtOrEtq230
+NuTtUAaJHIHNtv4CvpUcbtlFMWFYUUum7d22A8YTfUeccG5PsjjCoQG/dhhSbKNZ
+MFcwJgYDVR0RBB8wHYIbaGVscGVyMy5wcm9kLmlwYS1oZWxwZXIuZGV2MA4GA1Ud
+DwEB/wQEAwICpDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwCgYIKoZI
+zj0EAwIDSAAwRQIhAOTSQWbN7kfIatNJEwWTBL4xOY88E3+SOnBNExCsTkQuAiBB
+/cwOQQUEeE4llrDp+EnyGbzmVm5bINz8gePIxkKqog==
+-----END CERTIFICATE-----
+"""
+url = "helper3.org:443"
+
+[peers.hpke]
+public_key = "55f87a8794b4de9a60f8ede9ed000f5f10c028e22390922efc4fb63bc6be0a61"
+"#;
+
+    /// Valid: A sharded configuration
+    const SHARDED_OK: &str = r#"
+[[peers]]
+certificate = """
+-----BEGIN CERTIFICATE-----
+MIIBmzCCAUGgAwIBAgIIMlnveFys5QUwCgYIKoZIzj0EAwIwJjEkMCIGA1UEAwwb
+aGVscGVyMS5wcm9kLmlwYS1oZWxwZXIuZGV2MB4XDTI0MDkwNDAzMzMwM1oXDTI0
+MTIwNDAzMzMwM1owJjEkMCIGA1UEAwwbaGVscGVyMS5wcm9kLmlwYS1oZWxwZXIu
+ZGV2MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEWmrrkaKM7HQ0Y3ZGJtHB7vfG
+cT/hDCXCoob4pJ/fpPDMrqhiwTTck3bNOuzv9QIx+p5C2Qp8u67rYfK78w86NaNZ
+MFcwJgYDVR0RBB8wHYIbaGVscGVyMS5wcm9kLmlwYS1oZWxwZXIuZGV2MA4GA1Ud
+DwEB/wQEAwICpDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwCgYIKoZI
+zj0EAwIDSAAwRQIhAKVdDCQeXLRXDYXy4b1N1UxD/JPuD9H7zeRb8/nmIDTfAiBL
+a6L0t1Ug8i2RcequSo21x319Tvs5nUbGwzMFSS5wKA==
+-----END CERTIFICATE-----
+"""
+url = "helper1.shard0.org:443"
+shard_url = "helper1.shard0.org:444"
+
+[peers.hpke]
+public_key = "f458d5e1989b2b8f5dacd4143276aa81eaacf7449744ab1251ff667c43550756"
+
+[[peers]]
+certificate = """
+-----BEGIN CERTIFICATE-----
+MIIBmzCCAUGgAwIBAgIITOtoca16QckwCgYIKoZIzj0EAwIwJjEkMCIGA1UEAwwb
+aGVscGVyMi5wcm9kLmlwYS1oZWxwZXIuZGV2MB4XDTI0MDkwNDAzMzMwOFoXDTI0
+MTIwNDAzMzMwOFowJjEkMCIGA1UEAwwbaGVscGVyMi5wcm9kLmlwYS1oZWxwZXIu
+ZGV2MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAETxOH4ATz6kBxLuRznKDFRugm
+XKmH7mzRB9wn5vaVlVpDzf4nDHJ+TTzSS6Lb3YLsA7jrXDx+W7xPLGow1+9FNqNZ
+MFcwJgYDVR0RBB8wHYIbaGVscGVyMi5wcm9kLmlwYS1oZWxwZXIuZGV2MA4GA1Ud
+DwEB/wQEAwICpDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwCgYIKoZI
+zj0EAwIDSAAwRQIhAI4G5ICVm+v5KK5Y8WVetThtNCXGykUBAM1eE973FBOUAiAS
+XXgJe9q9hAfHf0puZbv0j0tGY3BiqCkJJaLvK7ba+g==
+-----END CERTIFICATE-----
+"""
+url = "helper2.shard0.org:443"
+shard_url = "helper2.shard0.org:555"
+
+[peers.hpke]
+public_key = "62357179868e5594372b801ddf282c8523806a868a2bff2685f66aa05ffd6c22"
+
+[[peers]]
+certificate = """
+-----BEGIN CERTIFICATE-----
+MIIBmzCCAUGgAwIBAgIIaf7eDCnXh2swCgYIKoZIzj0EAwIwJjEkMCIGA1UEAwwb
+aGVscGVyMy5wcm9kLmlwYS1oZWxwZXIuZGV2MB4XDTI0MDkwNDAzMzMxMloXDTI0
+MTIwNDAzMzMxMlowJjEkMCIGA1UEAwwbaGVscGVyMy5wcm9kLmlwYS1oZWxwZXIu
+ZGV2MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEIMqxCCtu4joFr8YtOrEtq230
+NuTtUAaJHIHNtv4CvpUcbtlFMWFYUUum7d22A8YTfUeccG5PsjjCoQG/dhhSbKNZ
+MFcwJgYDVR0RBB8wHYIbaGVscGVyMy5wcm9kLmlwYS1oZWxwZXIuZGV2MA4GA1Ud
+DwEB/wQEAwICpDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwCgYIKoZI
+zj0EAwIDSAAwRQIhAOTSQWbN7kfIatNJEwWTBL4xOY88E3+SOnBNExCsTkQuAiBB
+/cwOQQUEeE4llrDp+EnyGbzmVm5bINz8gePIxkKqog==
+-----END CERTIFICATE-----
+"""
+url = "helper3.shard0.org:443"
+shard_url = "helper3.shard0.org:666"
+
+[peers.hpke]
+public_key = "55f87a8794b4de9a60f8ede9ed000f5f10c028e22390922efc4fb63bc6be0a61"
+"#;
+
+    /// Valid: Three sharded configs together for 9
+    static SHARDED_OK_REPEAT: Lazy<String> = Lazy::new(|| {
+        format!(
+            "{}{}{}",
+            SHARDED_OK,
+            SHARDED_OK.replace("shard0", "shard1"),
+            SHARDED_OK.replace("shard0", "shard2")
+        )
+    });
+
+    /// Invalid: A network toml with 8 entries
+    static SHARDED_8: Lazy<String> = Lazy::new(|| {
+        let last_peers_index = SHARDED_OK_REPEAT.rfind("[[peers]]").unwrap();
+        SHARDED_OK_REPEAT[..last_peers_index].to_string()
+    });
+
+    /// Invalid: Same as [`SHARDED_OK_REPEAT`] but without the expected ports
+    static SHARDED_MISSING_URLS_REPEAT: Lazy<String> = Lazy::new(|| {
+        let lines: Vec<&str> = SHARDED_OK_REPEAT.lines().collect();
+        let new_lines: Vec<String> = lines
+            .iter()
+            .filter(|line| !line.starts_with("shard_url ="))
+            .map(std::string::ToString::to_string)
+            .collect();
+        new_lines.join("\n")
+    });
 }
